Cleaned up PersonReport Example code
